refactor: make appendLocationToHistory inline to the weather subsriber update
The linter had been throwing this warning for a while. I reflecting about it for awhile I decided that due to the timings the previousLocations list was always up to date when this triggered, but the linter doesn't know this. In future the timing could be changed and break this assumption. Hence I have moved the body of it into the effect, and utilised the callback form of the setter so that the location list is always up to date.

Additionally this reduces the debounce time to 1 second, to make it a little more responsive.